% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mapping.R
\docType{data}
\name{agnostic_dnase_summary_V4_eQTL}
\alias{agnostic_dnase_summary_V4_eQTL}
\title{Internal Datasets used in create_cellSTAAR_mapping_file function}
\usage{
agnostic_dnase_summary_V4_eQTL
}
\description{
Internal Datasets used in create_cellSTAAR_mapping_file function
}
\keyword{datasets}
