% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cellSTAAR_pvalue.R
\name{compute_cellSTAAR_pvalue}
\alias{compute_cellSTAAR_pvalue}
\title{compute_cellSTAAR_pvalue.}
\usage{
compute_cellSTAAR_pvalue(
  data_obj,
  grouping_vars = c("gene", "chr", "phenotype", "mapping", "class", "ct_name")
)
}
\arguments{
\item{data_obj}{data frame of all results from the \code{run_cellSTAAR}. By controlling the \code{grouping_vars} parameter, multiple phenotypes, cell types, linking types, and genes can be input simultaneously.}

\item{grouping_vars}{set of variables that uniquely identify a row in \code{data_obj}.}
}
\description{
compute_cellSTAAR_pvalue.
}
