% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cellSTAAR_mapping_file.R
\name{create_cellSTAAR_mapping_file}
\alias{create_cellSTAAR_mapping_file}
\title{create_cellSTAAR_mapping_file.}
\usage{
create_cellSTAAR_mapping_file(
  gds.path,
  atac_file_path,
  ct_name,
  num_ct_samples = NULL,
  chr,
  link_types_to_run,
  element_class,
  out_wd,
  ncores = 1,
  genes_manual,
  sc_cutoff = 0.8
)
}
\arguments{
\item{gds.path}{Path to the gds file.}

\item{atac_file_path}{File path to the ATAC-seq data files. It is expected that both .bw and .bed files will be in the same directory.}

\item{ct_name}{Name of the cell type.}

\item{num_ct_samples}{Number of samples ABOVE 1. Set to NULL if the cell type has one sample, otherwise set to the total number of samples. It is expected that the samples will have a similar file names: e.g. if \code{num_ct_samples=2} and \code{ct_name} is Hepatocyte, the files will have the name "Hepatocyte_1" and "Hepatocyte_2".}

\item{chr}{chromosome given as a numeric value from 1-22.}

\item{link_types_to_run}{Which link types to run. The function will loop over the link_types.}

\item{element_class}{One of the three ENCODE V3 cCRE categories: dELS, pELS, and PLS.}

\item{out_wd}{Directory to save the mapping files. It is assumed that within the directory there will be sub directories "chr1" through "chr22".}

\item{ncores}{Number of cores to use in \code{pblapply} call.}

\item{genes_manual}{Names of genes to manually run mapping files on. If NULL, all protein coding genes in the chromosome being run will be used.}

\item{sc_cutoff}{Internal argument used to adjust the level of filtering from sc ATAC-seq data. cellSTAAR by default uses 0.8.}
}
\value{
a list
##' \itemize{
}
}
\description{
create_cellSTAAR_mapping_file.
}
