% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cellSTAAR_mapping_file_eQTL.R
\name{create_cellSTAAR_mapping_file_eQTL}
\alias{create_cellSTAAR_mapping_file_eQTL}
\title{create_cellSTAAR_mapping_file.}
\usage{
create_cellSTAAR_mapping_file_eQTL(
  gds.path,
  eQTL_match_file,
  name_to_append = "eQTL",
  chr,
  element_class,
  link_types_to_run,
  out_wd,
  ncores = 1,
  genes_manual = NULL
)
}
\arguments{
\item{gds.path}{Path to the gds file.}

\item{eQTL_match_file}{match file for eQTL summary statistics.}

\item{name_to_append}{character to describe mapping file}

\item{chr}{chromosome given as a numeric value from 1-22.}

\item{element_class}{One of the three ENCODE V3 cCRE categories: dELS, pELS, and PLS. Users can run dELS and pELS in one function call, but PLS must be run separately.}

\item{link_types_to_run}{Character vector of link types to run. The function loops over all link types specified.}

\item{out_wd}{Directory to save the mapping files.}

\item{ncores}{Number of cores to use in \code{pblapply} call.}

\item{genes_manual}{Names of genes to manually run mapping files on. If NULL, all protein coding genes in the chromosome being run will be used. If specifying, ensure, the gene names used are proper HGNC symbols in the chromosome being computed.}
}
\value{
a sparse matrix, with rows covering variant positions and colnames covering protein coding genes. A value of 1 indicates a link between the respective position and the particular gene, 0 indicates no link.
}
\description{
create_cellSTAAR_mapping_file.
}
