% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cellSTAAR_mapping_file_multlink.R
\name{create_cellSTAAR_mapping_file_multlink}
\alias{create_cellSTAAR_mapping_file_multlink}
\title{create_cellSTAAR_mapping_file_multlink (EXPERIMENTAL, used to combine distance linking and non distance linking into to masks).}
\usage{
create_cellSTAAR_mapping_file_multlink(
  gds.path,
  sc_epi_file_path,
  ct_name,
  num_replicate_ct_samples = NULL,
  chr,
  element_class,
  link_types_to_run,
  cCRE_version = "V3",
  QC_label = "annotation/filter",
  out_wd,
  ncores = 1,
  genes_manual = NULL,
  sc_cutoff = 0.8
)
}
\arguments{
\item{gds.path}{Path to the gds file.}

\item{sc_epi_file_path}{File path to the ATAC-seq data files. It is expected that both .bw and .bed files will be in the same directory. In the cellSTAAR manuscript, these are scATAC-seq datasets from the CATlas resource as downloaded from https://decoder-genetics.wustl.edu/catlasv1/humanenhancer/data/bw/ and https://decoder-genetics.wustl.edu/catlasv1/humanenhancer/data/Peaks/.}

\item{ct_name}{Name of the cell type, used for (1) loading scATAC-seq data and (2) in the created file name. If "none" is specified, a mapping file consisting of all variants in an ENCODE V3 element class (as specified by \code{element_class}) which are linked using the specific linking approach will be output.}

\item{num_replicate_ct_samples}{Number of samples ABOVE 1. Set to NULL if the cell type has one sample, otherwise set to the total number of samples. It is expected that the samples will have similar file names: e.g. if \code{num_replicate_ct_samples=3} and \code{ct_name} is Hepatocyte, the files will have the name "Hepatocyte_1",  "Hepatocyte_2", and "Hepatocyte_3".}

\item{chr}{chromosome given as a numeric value from 1-22.}

\item{element_class}{One of the three ENCODE V3 cCRE categories: dELS, pELS, and PLS. Users can run dELS and pELS in one function call, but PLS must be run separately.}

\item{link_types_to_run}{Character vector of link types to run. The function loops over all link types specified.}

\item{cCRE_version}{Which version of ENCODE cCREs should be used to generate mapping files? Must be one of "V3" or "V4". Defaults to "V3".}

\item{QC_label}{Gives the position within the GDS file of the variable which flags whether a given variant passed quality. It is assumed that the variable will have values of "PASS" and "FAIL." Defaults to "annotation/filter". See the cellSTAAR README for more details.}

\item{out_wd}{Directory to save the mapping files.}

\item{ncores}{Number of cores to use in \code{pblapply} call.}

\item{genes_manual}{Names of genes to manually run mapping files on. If NULL, all protein coding genes in the chromosome being run will be used. If specifying, ensure, the gene names used are proper HGNC symbols in the chromosome being computed.}

\item{sc_cutoff}{Threshold of non-zero scores to filter CATLas bigwig files. Defaults to 0.8 (used in cellSTAAR manuscript analysis).}
}
\value{
a sparse matrix, with rows covering variant positions and colnames covering protein coding genes. A value of 1 indicates a link between the respective position and the particular gene, 0 indicates no link.
}
\description{
create_cellSTAAR_mapping_file_multlink (EXPERIMENTAL, used to combine distance linking and non distance linking into to masks).
}
