% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ct_annotations.R
\name{create_ct_annotations}
\alias{create_ct_annotations}
\title{create_ct_annotations.}
\usage{
create_ct_annotations(
  gds.path,
  sc_epi_file_path,
  ct_name,
  num_replicate_ct_samples,
  chr,
  QC_label = "annotation/filter",
  out_wd
)
}
\arguments{
\item{gds.path}{Path to the gds file.}

\item{sc_epi_file_path}{File path to the ATAC-seq data files. It is expected that both .bw and .bed files will be in the same directory.}

\item{ct_name}{Name of the cell type, used for (1) loading scATAC-seq data and (2) in the file name.}

\item{num_replicate_ct_samples}{Number of samples ABOVE 1. Set to NULL if the cell type has one sample, otherwise set to the total number of samples. It is expected that the samples will have similar file names: e.g. if \code{num_replicate_ct_samples=3} and \code{ct_name} is Hepatocyte, the files will have the name "Hepatocyte_1",  "Hepatocyte_2", and "Hepatocyte_3".}

\item{chr}{chromosome number (used as part of output filename).}

\item{QC_label}{Gives the position within the GDS file of the variable which flags whether a given variant passed quality. It is assumed that the variable will have values of "PASS" and "FAIL." Defaults to "annotation/filter". See the cellSTAAR README for more details.}

\item{out_wd}{Directory to save the aPCs files.}
}
\value{
a numeric vector of cell-type PHRED-scaled values (called "aPCs" for consistency within STAAR family)
}
\description{
create_ct_annotations.
}
