% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ct_aPCs.R
\name{create_ct_aPCs}
\alias{create_ct_aPCs}
\title{create_ct_aPCs.}
\usage{
create_ct_aPCs(gds.path, atac_file_path, ct_name, num_ct_samples, chr, out_wd)
}
\arguments{
\item{gds.path}{Path to the gds file.}

\item{atac_file_path}{File path to the ATAC-seq data files. It is expected that both .bw and .bed files will be in the same directory.}

\item{ct_name}{Name of the cell type, used for (1) loading scATAC-seq data and (2) in the file name.}

\item{num_ct_samples}{Number of samples ABOVE 1. Set to NULL if the cell type has one sample, otherwise set to the total number of samples. It is expected that the samples will have a similar file names: e.g. if \code{num_ct_samples=2} and \code{ct_name} is Hepatocyte, the files will have the name "Hepatocyte_1" and "Hepatocyte_2".}

\item{chr}{chromosome number (used as part of output filename).}

\item{out_wd}{Directory to save the mapping files. It is assumed that within the directory there will be sub directories "chr1" through "chr22".}
}
\value{
a numeric vector of cell-type PHRED-scaled values (called "aPCs" for consistency within STAAR family)
}
\description{
create_ct_aPCs.
}
