% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mapping.R
\docType{data}
\name{genes_biomaRt_all}
\alias{genes_biomaRt_all}
\title{Internal Datasets used in create_cellSTAAR_mapping_file function}
\usage{
genes_biomaRt_all
}
\description{
Internal Datasets used in create_cellSTAAR_mapping_file function
}
\keyword{datasets}
