% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mapping.R
\docType{data}
\name{raw_mappings_cCRE_V3_dist_50000_100000}
\alias{raw_mappings_cCRE_V3_dist_50000_100000}
\title{Internal Datasets used in create_cellSTAAR_mapping_file function}
\usage{
raw_mappings_cCRE_V3_dist_50000_100000
}
\description{
Internal Datasets used in create_cellSTAAR_mapping_file function
}
\keyword{datasets}
