% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mapping.R
\docType{data}
\name{raw_mappings_cCRE_V3_EpiMap_link_all_50}
\alias{raw_mappings_cCRE_V3_EpiMap_link_all_50}
\title{Internal Datasets used in create_cellSTAAR_mapping_file function}
\usage{
raw_mappings_cCRE_V3_EpiMap_link_all_50
}
\description{
Internal Datasets used in create_cellSTAAR_mapping_file function
}
\keyword{datasets}
