% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellSTAAR.R
\name{run_cellSTAAR}
\alias{run_cellSTAAR}
\title{runcellSTAAR.}
\usage{
run_cellSTAAR(
  gds.path,
  ct_names,
  mapping_object_list,
  chr,
  phenotype,
  ct_aPC_list,
  null_model,
  variants_to_condition_on = data.frame(),
  annotation_name_catalog,
  ncores_small = 1,
  ncores_large = 1,
  variables_to_add_to_output = NULL,
  chr.id = 1,
  n_splits = 1,
  genes_manual = NULL,
  gwas_cat_file_path = NULL,
  gwas_cat_vals = NULL,
  return_results = FALSE,
  save_results = TRUE,
  out_dir = "/",
  rare_maf_cutoff = 0.01
)
}
\arguments{
\item{gds.path}{Path to the gds file.}

\item{ct_names}{ct_names.}

\item{mapping_object_list}{An object of class 'list' with each element being a mapping file output from the \code{create_cellSTAAR_mapping_file} function. All objects should represent the the same link approach to have structured output.}

\item{chr}{chromosome given as a numeric value from 1-22.}

\item{phenotype}{Character name of the phenotype being analyzed. Provided as part of output.}

\item{ct_aPC_list}{An object of class 'list' with each element being a ct_aPC object output from the \code{create_cellSTAAR_ct_aPCs} function.}

\item{null_model}{null model}

\item{variants_to_condition_on}{Dataframe of conditional variants with columns}

\item{annotation_name_catalog}{Dataframe with column names and locations in the .gds file for the functional annotations to include.}

\item{ncores_small}{Number of cores for genes with small variant sets (<500 variants)}

\item{ncores_large}{Number of cores for genes with large variant sets (>500 variants)}

\item{variables_to_add_to_output}{Dataframe of additional variables to add to output.}

\item{chr.id}{Used to split a chromosome into multiple jobs. Must be <= the \code{n_splits} parameter.}

\item{n_splits}{Number of splits for the chomosome. Used to split genes.}

\item{genes_manual}{Names of genes to manually analyze. Otherwise, all protein coding genes in the chromosome are analyzed.}

\item{gwas_cat_file_path}{gwas catalog file path.}

\item{gwas_cat_vals}{Values from gwas catalog file.}

\item{return_results}{If \code{TRUE}, the dataframe of results will be returned.}

\item{save_results}{If \code{TRUE}, the dataframe of results saved in the \code{out_dir} directory.}

\item{out_dir}{Directory to save results (used only if \code{save_results} is TRUE).}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default = 0.01)}
}
\value{
A data frame with the following columns, and additionally any columns passed in through the \code{variables_to_add_to_output} parameter:
##' \itemize{
\item{\code{num_rare_SNV: }}{Number of SNV tested using \code{STAAR} function call. Equal to \code{NA} if no variants in variant set or conditional analysis run instead.}
\item{\code{pval_STAAR_O: }}{Omnibus p-value reported from \code{STAAR} function call. Equal to \code{NA} if no variants in variant set or conditional analysis run instead.}
\item{\code{num_individuals: }}{Number of individuals used in analysis.}
\item{\code{phenotype: }}{Phenotype analyzed.}
\item{\code{ct_name: }}{Name of the cell type used in variant set construction and functional annotation.}
\item{\code{gene: }}{Gene name.}
\item{\code{STAAR_time_taken: }}{Amount of time taken within \code{STAAR} function call. Equal to \code{NA} if no variants in variant set or conditional analysis run instead.}
##' \item{\code{num_rare_SNV_cond: }}{Number of SNV tested using \code{STAAR_cond} function call. Equal to \code{NA} if no variants in variant set or unconditional analysis run instead.}
\item{\code{pval_STAAR_O_cond: }}{Omnibus p-value reported from \code{STAAR_cond} function call. Equal to \code{NA} if no variants in variant set or unconditional analysis run instead.}
\item{\code{n_known_var_cond: }}{Number of variants conditioned on.}
\item{\code{rsIDs_cond: }}{rsIDs of variants conditioned on.}
\item{\code{STAAR_cond_time_taken: }}{Amount of time taken within \code{STAAR_cond} function call. Equal to \code{NA} if no variants in variant set or unconditional analysis run instead.}
\item{\code{date: }}{Date analysis was run.}
\item{\code{ncores_max: }}{Maximum number of cores used in analysis.}
\item{\code{job_time_taken: }}{Total time taken in cellSTAAR function call.}
\item{\code{chr: }}{Chromosome number.}
}
}
\description{
runcellSTAAR.
}
