% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cellSTAAR_pvalue.R
\name{compute_cellSTAAR_pvalue}
\alias{compute_cellSTAAR_pvalue}
\title{compute_cellSTAAR_pvalue.}
\usage{
compute_cellSTAAR_pvalue(
  data_obj,
  grouping_vars = c("gene", "chr", "phenotype", "element_class", "ct_name"),
  use_conditional_p_if_exist = FALSE
)
}
\arguments{
\item{data_obj}{Data frame of all results from the \code{run_cellSTAAR} function. By controlling the \code{grouping_vars} parameter, multiple phenotypes, cell types, linking types, and genes can be input simultaneously.}

\item{grouping_vars}{Set of variables (Other than "link_type"!) that uniquely identify a row in \code{data_obj}.}

\item{use_conditional_p_if_exist}{Should the function compute the cellSTAAR omnibus p-value using conditional p-values when they exist? Defaults to FALSE.}
}
\description{
compute_cellSTAAR_pvalue.
}
