% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellSTAAR.R
\name{run_cellSTAAR}
\alias{run_cellSTAAR}
\title{run_cellSTAAR.}
\usage{
run_cellSTAAR(
  gds.path,
  ct_names,
  chr,
  phenotype,
  mapping_object_list,
  element_class,
  link_type,
  ct_annotation_list,
  null_model,
  QC_label = "annotation/filter",
  variants_to_condition_on = data.frame(),
  annotation_name_catalog,
  analysis_to_run = "unconditional_only",
  ncores_small = 1,
  ncores_large = 1,
  variables_to_add_to_output = NULL,
  chr.id = 1,
  n_splits = 1,
  genes_manual = NULL,
  return_results = FALSE,
  save_results = TRUE,
  out_dir = "/",
  rare_maf_cutoff = 0.01,
  gwas_cat_file_path = NULL,
  gwas_cat_vals = NULL
)
}
\arguments{
\item{gds.path}{Path to the gds file.}

\item{ct_names}{Character vector of cell type names to run.}

\item{chr}{chromosome given as a numeric value from 1-22.}

\item{phenotype}{Character name of the phenotype being analyzed. Provided as part of output.}

\item{mapping_object_list}{An object of class 'list' with each element being a mapping file output from the \code{create_cellSTAAR_mapping_file} function. All objects should represent the the same link approach to have logical output, and it is assumed that the name of each item in the list is "map_obj_CTNAME", where all values in input \code{ct_names} have a corresponding item in \code{mapping_object_list}.}

\item{element_class}{One of the three ENCODE V3 cCRE categories: dELS, pELS, and PLS.}

\item{link_type}{Linking type corresponding to the objects in \code{mapping_object_list}.}

\item{ct_annotation_list}{An object of class 'list' with each element being an object output from the \code{create_cellSTAAR_ct_annotations} function.}

\item{null_model}{Null model object output from the \code{fit_null_glmmkin} function of the \code{STAAR} package.}

\item{QC_label}{Gives the position within the GDS file of the variable which flags whether a given variant passed quality. It is assumed that the variable will have values of "PASS" and "FAIL." Defaults to "annotation/filter". See the cellSTAAR README for more details.}

\item{variants_to_condition_on}{Data frame of variants to condition on. Expected to have columns "CHR", "POS", "REF", "ALT", "rsID", and "phenotype". Defaults to an empty data frame, meaning unconditional analysis will be run for all genes. If supplied, cellSTAAR will run conditional analysis using all variants in \code{variants_to_condition_on} within +- 1 Mega base.}

\item{annotation_name_catalog}{Data frame with column names and locations in the GDS file for the functional annotations to include.}

\item{analysis_to_run}{Options are "unconditional_only", "conditional_if_needed", or "both". If "conditional_if_needed", should specify \code{variants_to_condition_on}. Defaults to "unconditional_only" meaning no conditional analysis will be attempted. If "conditional_if_needed", only conditional analysis will be returned if there are known variants to condition on within +- 1 Mega base, otherwise only unconditional will be returned. If "both", conditional and unconditional will be attempted (note this can add substantial computation time if many genes require conditional analysis).}

\item{ncores_small}{Number of cores for genes with small variant sets (<500 variants)}

\item{ncores_large}{Number of cores for genes with large variant sets (>500 variants)}

\item{variables_to_add_to_output}{Data frame of one row with additional variables to add to output. Useful for strutured output to pass into the \code{compute_cellSTAAR_pvalue} function.}

\item{chr.id}{Used to split the genes from the analyzed chromosome into multiple jobs.. Must be <= the \code{n_splits} parameter. Defaults to 1, meaning the entire chromosome is analyzed in one job.}

\item{n_splits}{Total number of splits for genes from the chromosome being analyzed. Used to distribute computation across multiple function calls. Defaults to 1, meaning the entire chromosome is analyzed in one job.}

\item{genes_manual}{Names of genes to manually run mapping files on. If NULL (default), all protein coding genes in the chromosome (as taken from the mapping objects in \code{mapping_object_list}) will be used. If specifying, ensure the gene names used are proper HGNC symbols in the chromosome being analyzed.}

\item{return_results}{If \code{TRUE}, the data frame of results will be returned.}

\item{save_results}{If \code{TRUE}, the data frame of results will be saved in the \code{out_dir} directory.}

\item{out_dir}{Directory to save results (used only if \code{save_results} is TRUE).}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in defining rare variants (default = 0.01).}

\item{gwas_cat_file_path}{File path to a GWAS catalog file. This step is used to remove any rare variants that are contained within the GWAS catalog from the variant sets being tested. May be of interest if conditional analysis is not used.}

\item{gwas_cat_vals}{Terms from the GWAS catalog corresponding to the phenotype being analyzed.}
}
\value{
A data frame with the following columns, and additionally any columns passed in through the \code{variables_to_add_to_output} parameter:
##' \itemize{
\item{\code{num_rare_SNV: }}{Number of SNV tested using \code{STAAR} function call. Equal to \code{NA} if no variants in variant set or only conditional analysis run.}
\item{\code{pval_STAAR_O: }}{Omnibus p-value reported from \code{STAAR} function call. Equal to \code{NA} if no variants in variant set or only conditional analysis run.}
\item{\code{num_individuals: }}{Number of individuals used in analysis.}
\item{\code{phenotype: }}{Phenotype analyzed.}
\item{\code{ct_name: }}{Name of the cell type used in variant set construction and functional annotation.}
\item{\code{gene: }}{Gene name.}
\item{\code{STAAR_time_taken: }}{Amount of time taken within \code{STAAR} function call. Equal to \code{NA} if no variants in variant set or conditional analysis run instead.}
##' \item{\code{num_rare_SNV_cond: }}{Number of SNV tested using \code{STAAR_cond} function call. Equal to \code{NA} if no variants in variant set, no variants to condition on within 1 Mb, or only unconditional analysis run instead.}
\item{\code{pval_STAAR_O_cond: }}{Omnibus p-value reported from \code{STAAR_cond} function call. Equal to \code{NA} if no variants in variant set, no variants to condition on within 1 Mb, or only unconditional analysis run instead.}
\item{\code{n_known_var_cond: }}{Number of variants conditioned on. Equal to \code{NA} if no variants in variant set, no variants to condition on within 1 Mb, or only unconditional analysis run instead.}
\item{\code{rsIDs_cond: }}{rsIDs of variants conditioned on. Equal to \code{NA} if no variants in variant set, no variants to condition on within 1 Mb, or only unconditional analysis run instead.}
\item{\code{STAAR_cond_time_taken: }}{Amount of time taken within \code{STAAR_cond} function call. Equal to \code{NA} if no variants in variant set or unconditional analysis run instead. }
\item{\code{date: }}{Date analysis was run.}
\item{\code{ncores_max: }}{Maximum number of cores used in analysis.}
\item{\code{job_time_taken: }}{Total time taken in cellSTAAR function call.}
\item{\code{chr: }}{Chromosome number.}
}
}
\description{
run_cellSTAAR.
}
